requirements.txt
pygame-ce
pygame_gui
numpy
python-dotenv
pymunk
pillow